renamed constants
