+++ src/core/constants/ids/button/intex.ts
@@ -3,3 +3,5 @@ export const CDEP_BUTTON_CRIAR_CONTA = 'CDEP_BUTTON_CRIAR_CONTA';
 export const CDEP_BUTTON_CADASTRAR = 'CDEP_BUTTON_CADASTRAR';
 export const CDEP_BUTTON_CANCELAR = 'CDEP_BUTTON_CANCELAR';
 export const CDEP_BUTTON_VOLTAR = 'CDEP_BUTTON_VOLTAR';
+export const CDEP_BUTTON_MODAL_ALTERAR = 'CDEP_BUTTON_MODAL_ALTERAR';
+export const CDEP_BUTTON_MODAL_CANCELAR = 'CDEP_BUTTON_MODAL_CANCELAR';
