+++ src/core/constants/lista-tipo-usuario.ts
