+++ src/core/dto/perfil-usuario-dto.ts
@@ -3,4 +3,4 @@ import { Guid } from '@microsoft/sp-core-library';
 export type PerfilUsuarioDTO = {
   perfil: Guid;
   perfilNome: string;
-}
+};
