+++ src/pages/login/index.tsx
@@ -9,13 +9,13 @@ import autenticacaoService from '~/core/services/autenticacao-service';
 import { AxiosError } from 'axios';
 import { useNavigate } from 'react-router-dom';
 import ErroGeralLogin from '~/components/cdep/erro-geral-login';
-import { CDEP_BUTTON_ACESSAR, CDEP_BUTTON_CRIAR_CONTA } from '~/core/constats/ids/button/intex';
-import { CDEP_INPUT_LOGIN, CDEP_INPUT_SENHA } from '~/core/constats/ids/input';
+import { CDEP_BUTTON_ACESSAR, CDEP_BUTTON_CRIAR_CONTA } from '~/core/constants/ids/button/intex';
+import { CDEP_INPUT_LOGIN, CDEP_INPUT_SENHA } from '~/core/constants/ids/input';
 import {
   ERRO_INFORMAR_USUARIO_SENHA,
   ERRO_LOGIN,
   ERRO_LOGIN_SENHA_INCORRETOS,
-} from '~/core/constats/mensagens';
+} from '~/core/constants/mensagens';
 import { AutenticacaoDTO } from '~/core/dto/autenticacao-dto';
 import { RetornoBaseDTO } from '~/core/dto/retorno-base-dto';
 import { ValidateErrorEntity } from '~/core/dto/validate-error-entity';
