+++ src/pages/meus-dados/components/modal-edit-telefone/modal-edit-telefone.tsx
@@ -2,7 +2,7 @@ import { Form } from 'antd';
 import { useForm } from 'antd/es/form/Form';
 import React from 'react';
 import InputTelefone from '~/components/cdep/input/telefone';
-import { CDEP_INPUT_TELEFONE } from '~/core/constats/ids/input';
+import { CDEP_INPUT_TELEFONE } from '~/core/constants/ids/input';
 import { useAppSelector } from '~/core/hooks/use-redux';
 import usuarioService from '~/core/services/usuario-service';
 import ModalEditDefault from '../modal-edit-default';
